+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add map.getCameraTargetElevation() (#1558)
 - Add `freezeElevation` to `AnimationOptions` to allow smooth camera movement in 3D (#1514, #1492)
 - [Breaking] Remove deprecated mapboxgl css classes (#1575)
+- Add map.setStyle's transformStyle option (#1632)
 
 ### 🐞 Bug fixes